+++ zarrs_data_type/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Implement `From<num::complex::Complex<half::{bf16,f16}>>` for `FillValue`
+
 ## [0.3.0] - 2025-05-16
 
 ### Changed